+++ lib/src/types/entity.dart
@@ -343,7 +343,7 @@ class AssetEntity {
   final String? title;
 
   ///  * Android: `MediaStore.MediaColumns.DISPLAY_NAME`.
-  ///  * iOS/macOS: [`PHAssetResource.originalFilename`](https://developer.apple.com/documentation/photokit/phassetresource/1623985-originalfilename?language=objc).
+  ///  * iOS/macOS: `PHAssetResource.originalFilename`.
   Future<String> get titleAsync => plugin.getTitleAsync(this);
 
   /// {@macro photo_manager.AssetType}