+++ ios/Classes/core/PHAsset+PM_COMMON.h
@@ -18,6 +18,9 @@ NS_ASSUME_NONNULL_BEGIN
 - (int)unwrappedSubtype;
 
 - (NSString*)title;
+
+- (NSString *)originalFilenameWithSubtype:(int) subtype;
+
 /**
  Get the MIME type for this asset from UTI (`PHAssetResource.uniformTypeIdentifier`), such as `image/jpeg`, `image/heic`, `video/quicktime`, etc.