+++ ios/Classes/core/PHAsset+PM_COMMON.m
@@ -94,6 +94,25 @@
         return @"";
     }
 }
+
+- (NSString *)originalFilenameWithSubtype:(int)subtype {
+    if (@available(iOS 9.1, *)) {
+        if ([self isLivePhoto] && subtype == PHAssetMediaSubtypePhotoLive) {
+            return [self getLivePhotosResource].originalFilename;
+        }
+    }
+    if (@available(macOS 10.11, *)) {
+        if ([self isLivePhoto] && subtype == PHAssetMediaSubtypePhotoLive) {
+            return [self getLivePhotosResource].originalFilename;
+        }
+    }
+    PHAssetResource *resource = [self getAdjustResource];
+    if (resource) {
+        return resource.originalFilename;
+    }
+    return @"";
+}
+
 // UTI: https://developer.apple.com/library/archive/documentation/FileManagement/Conceptual/understanding_utis/understand_utis_intro/understand_utis_intro.html#//apple_ref/doc/uid/TP40001319
 - (NSString *)mimeType {
     PHAssetResource *resource = [[PHAssetResource assetResourcesForAsset:self] firstObject];